New Version

+++ Snake/src/wrapper/model/SnakeInterface.java
@@ -11,7 +11,8 @@ package wrapper.model;
 public interface SnakeInterface {
 	
 	    double snakeSpeed = 2;
-	    double radius = 13;
+	    
+	    int radius = 13;
 
 	    int GAME_WIDTH = 800;
